Quick fix to use Sharp Icons
--Assuming you have set up Pro--
Copy the contents of icon-families.json into the lib/fonts/icons.json

This code is backwards compatible so that if you use icons.json it will pull out the relevant fields.
icon-families.json doesn't have the styles tag so instead we can iterate over the svgs keys of "classic" and check if it has "sharp"
This also means that the key for sharp is now "sharp <thin, light, regular, solid> name"
so we have to do some manipulation to remove the empty space to map the Class names correctly and to un-comment out the relevant pubspec.yaml files.

I hope this helps someone out to start using Sharp Icons in this.